// src/types/article-search.ts
/**
 * This file was auto-generated by openapi-typescript.
 */

export interface Response {
  status: string
  copyright: string
  response: {
    docs: Definitions["Article"][]
    meta: {
      hits: number
      offset: number
      time: number
    }
  }
}

export interface Definitions {
  Article: {
    abstract: string
    lead_paragraph: string
    /** @description Article URL. */
    web_url: string
    snippet: string
    /** @description Page in print (e.g. 1). */
    print_page: number
    /** @description Section in print (e.g. A). */
    print_section: string
    source: string
    multimedia: Definitions["Multimedia"][]
    headline: Definitions["Headline"]
    keywords: Definitions["Keyword"][]
    /** @description Publication date. */
    pub_date: string
    /** @description Document type (article, multimedia). */
    document_type: string
    /** @description Desk in the newsroom that worked on the story (Foreign, Metro, Sports, ...). */
    news_desk: string
    /** @description Section that the article appeared in (New York, Sports, World, ...). */
    section_name: string
    byline: Definitions["Byline"]
    /** @description Type of asset (Correction, News, Op-Ed, Review, Video, ...). */
    type_of_material: string
    _id: string
    /** @description Number of words in the article. */
    word_count: number
    /** @description Uniquely identifies an asset. */
    uri: string
  }
  Byline: {
    original: string
    person: Definitions["Person"][]
    organization: string
  }
  Headline: {
    main: string
    kicker: string
    content_kicker: string
    print_headline: string
    name: string
    seo: string
    sub: string
  }
  Keyword: {
    name: string
    value: string
    rank: number
    major: string
  }
  Multimedia: {
    rank: number
    subtype: string
    caption: string
    credit: string
    type: string
    url: string
    height: number
    width: number
    legacy: {
      xlarge: string
      xlargewidth: number
      xlargeheight: number
    }
    crop_name: string
  }
  Person: {
    firstname: string
    middlename: string
    lastname: string
    qualifier: string
    title: string
    role: string
    organization: string
    rank: number
  }
}